Blocker Sonar Fix - XACMLPapServlet.java
Blocker Sonar Fix - XACMLPapServlet.java

Issue-ID: POLICY-1435

+++ ONAP-PAP-REST/src/main/java/org/onap/policy/pap/xacml/rest/XACMLPapServlet.java
@@ -1743,7 +1743,9 @@ public class XACMLPapServlet extends HttpServlet implements StdItemSetChangeList
                         }
 
                         //delete temporary policy file from the bin directory
-                        Files.deleteIfExists(Paths.get(policy.getId()));
+                        if(null != policy) {
+                            Files.deleteIfExists(Paths.get(policy.getId()));
+                        }
 
                     } catch (Exception e) {
                         acPutTransaction.rollbackTransaction();
